backend/requirements.txt
Django>=5.0,<6
psycopg2-binary
djangorestframework>=3.14,<4
django-cors-headers>=4.4,<5
django-filter>=24.2
drf-spectacular>=0.27,<0.28
